backend/requirements.txt
# FastAPI and dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0

# Database and storage
supabase==2.3.1

# ML and AI
transformers==4.36.2
torch==2.6.0
accelerate==0.24.1
huggingface_hub==0.20.2
openai==1.12.0

# File processing
python-docx==0.8.11

# Utilities
python-dotenv==1.0.0
uuid==1.30
python-dateutil==2.8.2
regex==2023.12.25